requirements.txt
customtkinter>=5.2.0
mysql-connector-python>=8.0.0
bcrypt>=4.0.0
